Combine relevant functions in apollo-client.ts
