+++ src/api/apollo-client.ts
@@ -6,9 +6,55 @@ import {
   ServerError,
 } from '@apollo/client'
 import { onError } from '@apollo/client/link/error'
-import { refreshAccessToken } from 'api/index'
+import axios from 'axios'
 import { toast } from 'react-hot-toast'
 import { forceLogOut } from 'utils/auth'
+
+const setItemWithExpiry = (key, value, ttl) =>
+  localStorage.setItem(
+    key,
+    JSON.stringify({
+      value,
+      expiry: new Date().getTime() + ttl,
+    }),
+  )
+
+const getItemWithExpiry = (key) => {
+  const itemStr = localStorage.getItem(key)
+
+  if (!itemStr) {
+    return null
+  }
+
+  const item = JSON.parse(itemStr)
+
+  if (new Date().getTime() > item.expiry) {
+    localStorage.removeItem(key)
+    return null
+  }
+
+  return item.value
+}
+
+const refreshAccessToken = async () => {
+  if (getItemWithExpiry('hvg:refreshingAccessToken') === 'true') {
+    // bail if we're already refreshing
+    return
+  }
+
+  try {
+    setItemWithExpiry('hvg:refreshingAccessToken', 'true', 10000)
+    await axios.post('/login/refresh', null, {
+      headers: {
+        accept: 'application/json',
+        'content-type': 'application/json',
+      },
+      withCredentials: true,
+    })
+  } finally {
+    localStorage.removeItem('hvg:refreshingAccessToken')
+  }
+}
 
 export const apolloClient = (() => {
   if (typeof window === 'undefined') {
